requirements.txt
openai-whisper
gradio==6.5.1
